+++ scripts/build-deb.sh
@@ -52,12 +52,12 @@ FPM="fpm --input-type dir --chdir $PACKAGE_TMPDIR --name yarn --version $VERSION
   `"--url https://yarnpkg.com/ --license BSD --description '$(cat resources/debian/description)'"
 
 ##### Build RPM (CentOS, Fedora) package
-./scripts/set-installation-method.js $PACKAGE_TMPDIR_ABSOLUTE/usr/share/yarn/package.json rpm
+./scripts/update-dist-manifest.js $PACKAGE_TMPDIR_ABSOLUTE/usr/share/yarn/package.json rpm
 eval "$FPM --output-type rpm  --architecture noarch --depends nodejs --category 'Development/Languages' ."
 mv *.rpm $OUTPUT_DIR
 
 ##### Build DEB (Debian, Ubuntu) package
-./scripts/set-installation-method.js $PACKAGE_TMPDIR_ABSOLUTE/usr/share/yarn/package.json deb
+./scripts/update-dist-manifest.js $PACKAGE_TMPDIR_ABSOLUTE/usr/share/yarn/package.json deb
 mkdir -p $PACKAGE_TMPDIR/DEBIAN
 mkdir -p $PACKAGE_TMPDIR/usr/share/lintian/overrides/
 cp resources/debian/lintian-overrides $PACKAGE_TMPDIR/usr/share/lintian/overrides/yarn
